Build Caddy server with Go 1.18 to fix OSX bug.
It appears to me that Go 1.18 has rewritten a bunch of code in its X509
certificate verification library to use different OSX functions that do
a better job. This fixes a bug I was encountering in which Caddy would
add a local root CA to the System keychain but could never verify that
it was there later, causing it to add it again every time the server is
restarted.