cargo: add curl-sys build script override to hooks
This will force projects that use these hooks and depend on curl-sys to
link against curl which we will provide in buildInputs.

We'll no longer silently statically compile against a vendored libcurl,
and we won't need pkg-config and curl.dev in nativeBuildInputs on Linux
(unless pkg-config is used for something else).

This can perhaps also be used for some under -sys libraries too if their
build scripts only handle linking to system libraries.